app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic>=1.9.0,<2.0.0
pymongo[srv]==3.12.0
python-dotenv==0.19.0
openai==1.43.1
python-multipart    